pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64"]

[dependency-groups]
dev = [
    "hypotest[dev]",
]

[project]
authors = [
    {email = "support@edisonscientific.com", name = "Edison Scientific Technical Staff"},
]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python",
    "Topic :: Software Development",
]
dependencies = [
    "aiodocker",
    "fhaviary[server,notebook]>=0.31.0",  # Pin for Message.is_multimodal fix
    "httpx>=0.27",
    "jupyter-client",
    "nbformat",
    "numpy",
    "pillow",
    "pydantic",
    "tenacity>=8.0",
]
description = "hypotest package"
keywords = []
maintainers = [
    {email = "support@edisonscientific.com", name = "Edison Scientific Technical Staff"},
]
name = "hypotest"
readme = "README.md"
requires-python = ">=3.11,<3.14"
version = "0.0.1"

[project.optional-dependencies]
dev = [
    "fhlmi>=0.25.0",  # Pin for headers
    "matplotlib",  # For tests that generate plots
    "mypy>=1.8",
    "prek",
    "pytest-asyncio",
    "pytest-recording",
    "pytest-rerunfailures",
    "pytest-subtests",  # TODO remove after pytest v9
    "pytest-sugar",
    "pytest-timeout",
    "pytest-timer[colorama]",
    "pytest-xdist",
    "pytest>=8",  # Pin to keep recent
    "types-Flask-SocketIO>=5.3.0",
    "types-PyYAML",
    "vcrpy>=8",  # Pin for dropping unused requests support
]

[tool.codespell]
check-filenames = true
check-hidden = true
skip = "uv.lock"

[tool.coverage]

[tool.coverage.report]
exclude_also = [
    "@overload",
    "if TYPE_CHECKING:",
]
precision = 2

[tool.coverage.run]
branch = true
command_line = "-m pytest"

[tool.mypy]
check_untyped_defs = true
enable_error_code = [
    "ignore-without-code",
    "mutable-override",
    "redundant-cast",
    "redundant-expr",
    "redundant-self",
    "truthy-bool",
    "truthy-iterable",
    "unimported-reveal",
    "unreachable",
    "unused-ignore",
]
error_summary = false
exclude = [
    "^\\.?venv",
]
platform = "linux"
plugins = ["pydantic.mypy"]
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pylint]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.main]
ignore-patterns = [
    "^version\\.py$",
]
jobs = 0

[tool.pylint.messages_control]
disable = [
    "bare-except",
    "broad-exception-caught",
    "broad-exception-raised",
    "consider-using-in",
    "cyclic-import",
    "dangerous-default-value",
    "empty-docstring",
    "expression-not-assigned",
    "fixme",
    "function-redefined",
    "global-statement",
    "global-variable-not-assigned",
    "import-outside-toplevel",
    "invalid-name",
    "keyword-arg-before-vararg",
    "line-too-long",
    "logging-fstring-interpolation",
    "logging-too-many-args",
    "missing-docstring",
    "missing-final-newline",
    "no-else-return",
    "no-member",
    "no-value-for-parameter",
    "protected-access",
    "raise-missing-from",
    "redefined-builtin",
    "too-few-public-methods",
    "too-many-ancestors",
    "too-many-arguments",
    "too-many-boolean-expressions",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-many-lines",
    "too-many-locals",
    "too-many-positional-arguments",
    "too-many-public-methods",
    "too-many-return-statements",
    "too-many-statements",
    "ungrouped-imports",
    "unidiomatic-typecheck",
    "unnecessary-dict-index-lookup",
    "unreachable",
    "unspecified-encoding",
    "unsubscriptable-object",
    "unsupported-membership-test",
    "unused-argument",
    "unused-import",
    "unused-variable",
    "use-sequence-for-iteration",
    "wrong-import-order",
    "wrong-import-position",
]
enable = [
    "useless-suppression",
]

[tool.pylint.reports]
score = false

[tool.pylint.similarities]
min-similarity-lines = 20

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
filterwarnings = []
testpaths = ["tests"]
timeout = 1000

[tool.refurb]
enable_all = true
ignore = [
    "FURB101",
    "FURB103",
    "FURB108",
    "FURB120",
    "FURB141",
    "FURB144",
    "FURB146",
    "FURB147",
    "FURB150",
    "FURB155",
]

[tool.ruff]
line-length = 120
unsafe-fixes = true

[tool.ruff.format]
docstring-code-format = true
preview = true

[tool.ruff.lint]
explicit-preview-rules = true
extend-select = [
    "ASYNC212",
    "ASYNC250",
    "B901",
    "B903",
    "B909",
    "B912",
    "CPY001",
    "DOC102",
    "DOC201",
    "DOC202",
    "DOC402",
    "DOC403",
    "DOC501",
    "DOC502",
    "E111",
    "E112",
    "E113",
    "E114",
    "E115",
    "E116",
    "E117",
    "E201",
    "E202",
    "E203",
    "E204",
    "E211",
    "E221",
    "E222",
    "E223",
    "E224",
    "E225",
    "E226",
    "E227",
    "E228",
    "E231",
    "E241",
    "E242",
    "E251",
    "E252",
    "E261",
    "E262",
    "E265",
    "E266",
    "E271",
    "E272",
    "E273",
    "E274",
    "E275",
    "E301",
    "E302",
    "E303",
    "E304",
    "E305",
    "E306",
    "E502",
    "FURB101",
    "FURB103",
    "FURB110",
    "FURB113",
    "FURB118",
    "FURB131",
    "FURB140",
    "FURB142",
    "FURB145",
    "FURB148",
    "FURB152",
    "FURB154",
    "FURB156",
    "FURB164",
    "FURB171",
    "FURB180",
    "FURB189",
    "FURB192",
    "LOG004",
    "PLC0207",
    "PLC1901",
    "PLC2701",
    "PLC2801",
    "PLE0304",
    "PLE1141",
    "PLE4703",
    "PLR0202",
    "PLR0203",
    "PLR0904",
    "PLR0914",
    "PLR0916",
    "PLR0917",
    "PLR1702",
    "PLR6104",
    "PLR6201",
    "PLR6301",
    "PLW0108",
    "PLW0244",
    "PLW1514",
    "PLW3201",
    "PT029",
    "RUF027",
    "RUF029",
    "RUF031",
    "RUF036",
    "RUF037",
    "RUF038",
    "RUF039",
    "RUF045",
    "RUF047",
    "RUF052",
    "RUF054",
    "RUF055",
    "RUF056",
    "RUF060",
    "RUF061",
    "RUF063",
    "RUF064",
    "RUF065",
    "RUF102",
    "TC008",
    "UP042",
    "W391",
]
external = [
    "FURB",
]
ignore = [
    "ANN",
    "ASYNC240",
    "BLE001",
    "C901",
    "COM812",
    "CPY001",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D203",
    "D212",
    "D402",
    "D406",
    "D407",
    "D413",
    "D415",
    "DOC201",
    "DOC402",
    "DOC501",
    "DOC502",
    "DTZ",
    "EM",
    "ERA001",
    "FBT001",
    "FBT002",
    "FIX",
    "FLY002",
    "G004",
    "INP001",
    "ISC001",
    "N803",
    "N806",
    "PLR0913",
    "PLR0917",
    "PLR1702",
    "PLR2004",
    "PLR6301",
    "PTH",
    "RUF027",
    "S101",
    "S105",
    "S311",
    "SLF001",
    "T201",
    "TC001",
    "TC002",
    "TC003",
    "TC006",
    "TD002",
    "TD003",
    "TID252",
    "TRY003",
    "W505",
]
preview = true
select = ["ALL"]
unfixable = [
    "B007",
    "B905",
    "ERA001",
    "F401",
    "F841",
    "PIE794",
    "RUF059",
    "TC004",
]

[tool.ruff.lint.flake8-annotations]
mypy-init-return = true

[tool.ruff.lint.per-file-ignores]
"**/prompts.py" = [
    "E501",
]
"**/tests/*.py" = [
    "N802",
    "PLR0914",
    "PLR1702",
    "PLR2004",
    "PLR6301",
]

[tool.ruff.lint.pycodestyle]
max-doc-length = 120

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.setuptools.packages.find]
where = ["src"]

[tool.tomlsort]
all = true
in_place = true
spaces_before_inline_comment = 2
spaces_indent_inline_array = 4
trailing_comma_inline_array = true

[tool.uv.sources]
hypotest = {workspace = true}
